+++ libs/shared/src/lib/services/index.ts
@@ -1,4 +1,6 @@
 export { AuthenticationService } from './authentication.service';
+export { ExternalContentService } from './external-content.service';
+export { ExternalUrlResolverService } from './external-url-resolver.service';
 export { NotificationsService } from './notifications.service';
 export { OperationsHistoryService } from './operations-history.service';
 export { ServicesHistoryService } from './services-history.service';